update floorfilter micro app
